Refactored InstructorController and SubmissionController logic to separate functionality. Implemented export() method so data from graded assignments is exported to a csv file.
