+++ Inspector.iml
@@ -32,5 +32,6 @@
       </library>
     </orderEntry>
     <orderEntry type="library" exported="" name="javafx-sdk-17.0" level="project" />
+    <orderEntry type="library" name="Maven: org.apache.commons:commons-csv:1.9.0" level="project" />
   </component>
 </module>
