+++ src/main/resources/views/SubmissionView.fxml
@@ -1,0 +1,120 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.TextArea?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.control.TreeView?>
+<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.layout.VBox?>
+<?import javafx.scene.text.Text?>
+
+<HBox maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" stylesheets="@../css/styles.css" xmlns="http://javafx.com/javafx/17" xmlns:fx="http://javafx.com/fxml/1" fx:controller="controllers.SubmissionController">
+   <VBox prefHeight="400.0" prefWidth="243.0" stylesheets="@../css/styles.css">
+      <Label id="username" fx:id="username">
+         <VBox.margin>
+            <Insets bottom="10.0" left="10.0" right="10.0" top="10.0" />
+         </VBox.margin>
+      </Label>
+      <TreeView id="treeView" fx:id="treeView" minHeight="400.0" prefWidth="200.0" />
+      <VBox alignment="TOP_CENTER" prefHeight="200.0" prefWidth="100.0">
+         <padding>
+            <Insets bottom="10.0" left="10.0" right="10.0" top="10.0" />
+         </padding>
+         <Label text="Grading Rubric" />
+         <HBox prefHeight="100.0" prefWidth="200.0" spacing="10.0">
+            <padding>
+               <Insets bottom="10.0" left="10.0" right="10.0" top="10.0" />
+            </padding>
+            <TextField id="criterion1MarkInput" fx:id="criterion1MarkInput" prefHeight="27.0" prefWidth="34.0" />
+            <TextField id="criterion1NameInput" fx:id="criterion1NameInput" />
+         </HBox>
+         <HBox layoutX="10.0" layoutY="10.0" prefHeight="100.0" prefWidth="200.0" spacing="10.0">
+            <padding>
+               <Insets bottom="10.0" left="10.0" right="10.0" top="10.0" />
+            </padding>
+            <TextField id="criterion2MarkInput" fx:id="criterion2MarkInput" prefHeight="27.0" prefWidth="34.0" />
+            <TextField id="criterion2NameInput" fx:id="criterion2NameInput" />
+         </HBox>
+         <HBox layoutX="10.0" layoutY="58.0" prefHeight="100.0" prefWidth="200.0" spacing="10.0">
+            <padding>
+               <Insets bottom="10.0" left="10.0" right="10.0" top="10.0" />
+            </padding>
+            <TextField id="criterion3MarkInput" fx:id="criterion3MarkInput" prefHeight="27.0" prefWidth="34.0" />
+            <TextField id="criterion3NameInput" fx:id="criterion3NameInput" />
+         </HBox>
+         <HBox layoutX="10.0" layoutY="106.0" prefHeight="100.0" prefWidth="200.0" spacing="10.0">
+            <padding>
+               <Insets bottom="10.0" left="10.0" right="10.0" top="10.0" />
+            </padding>
+            <TextField id="criterion4MarkInput" fx:id="criterion4MarkInput" prefHeight="27.0" prefWidth="34.0" />
+            <TextField id="criterion4NameInput" fx:id="criterion4NameInput" />
+         </HBox>
+         <Button mnemonicParsing="false" onAction="#setGradingRubric" text="Set Rubric" />
+      </VBox>
+   </VBox>
+   <VBox minWidth="600.0" prefWidth="950.0" styleClass="submission-display-section" HBox.hgrow="ALWAYS">
+      <HBox.margin>
+         <Insets bottom="20.0" left="20.0" right="20.0" top="20.0" />
+      </HBox.margin>
+      <Text id="submissionDisplay" fx:id="submissionDisplay" strokeType="OUTSIDE" strokeWidth="0.0" styleClass="root" wrappingWidth="378.13671875" />
+   </VBox>
+   <VBox alignment="TOP_CENTER" prefHeight="400.0" prefWidth="259.0">
+      <VBox>
+         <Label text="Marks" />
+         <HBox alignment="CENTER_LEFT" minWidth="300.0">
+            <padding>
+               <Insets bottom="10.0" left="10.0" right="10.0" top="10.0" />
+            </padding>
+            <TextField id="criterion1Mark" fx:id="criterion1Mark" onMouseExited="#updateTotalMarks" prefHeight="27.0" prefWidth="26.0" />
+            <Label id="criterion1Name" fx:id="criterion1Name" text="criterion 1" />
+         </HBox>
+         <HBox alignment="CENTER_LEFT">
+            <padding>
+               <Insets bottom="10.0" left="10.0" right="10.0" top="10.0" />
+            </padding>
+            <TextField id="criterion2Mark" fx:id="criterion2Mark" onMouseExited="#updateTotalMarks" prefHeight="27.0" prefWidth="26.0" />
+            <Label id="criterion2Name" fx:id="criterion2Name" text="criterion 2" />
+         </HBox>
+         <HBox alignment="CENTER_LEFT">
+            <padding>
+               <Insets bottom="10.0" left="10.0" right="10.0" top="10.0" />
+            </padding>
+            <TextField id="criterion3Mark" fx:id="criterion3Mark" onMouseExited="#updateTotalMarks" prefHeight="27.0" prefWidth="26.0" />
+            <Label id="criterion3Name" fx:id="criterion3Name" text="criterion 3" />
+         </HBox>
+         <HBox alignment="CENTER_LEFT">
+            <padding>
+               <Insets bottom="10.0" left="10.0" right="10.0" top="10.0" />
+            </padding>
+            <TextField id="criterion4Mark" fx:id="criterion4Mark" onMouseExited="#updateTotalMarks" prefHeight="27.0" prefWidth="26.0" />
+            <Label id="criterion4Name" fx:id="criterion4Name" text="criterion 4" />
+         </HBox>
+         <HBox alignment="CENTER_LEFT" prefHeight="43.0" prefWidth="262.0" spacing="10.0">
+            <padding>
+               <Insets bottom="10.0" left="10.0" right="10.0" top="10.0" />
+            </padding>
+            <Label id="maxMarksLabel" fx:id="maxMarksLabel" text="Max marks: " />
+         </HBox>
+         <HBox alignment="CENTER_LEFT" spacing="10.0">
+            <padding>
+               <Insets bottom="10.0" left="10.0" right="10.0" top="10.0" />
+            </padding>
+            <Label id="marksReceivedLabel" fx:id="marksReceivedLabel" text="Total marks: " />
+         </HBox>
+      </VBox>
+      <VBox alignment="CENTER" prefHeight="200.0" prefWidth="100.0" spacing="10.0">
+         <VBox.margin>
+            <Insets bottom="10.0" left="10.0" right="10.0" top="10.0" />
+         </VBox.margin>
+         <padding>
+            <Insets bottom="10.0" left="10.0" right="10.0" top="10.0" />
+         </padding>
+         <Label text="Feedback" />
+         <TextArea id="feedbackTextArea" fx:id="feedbackTextArea" prefHeight="400.0" prefWidth="200.0" wrapText="true" />
+         <Button mnemonicParsing="false" onAction="#saveMarks" text="Save" />
+      </VBox>
+      <Button id="exportButton" fx:id="exportButton" mnemonicParsing="false" onAction="#export" text="Export" />
+   </VBox>
+</HBox>
